Moved last block inserted reducer test from editor to the block-editor
