+++ packages/editor/src/store/test/reducer.native.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { postTitle, lastBlockInserted } from '../reducer';
+import { postTitle } from '../reducer';
 
 describe( 'state native', () => {
 	describe( 'postTitle', () => {
@@ -30,29 +30,6 @@ describe( 'state native', () => {
 				expect(
 					postTitle( { isSelected: false }, action ).isSelected
 				).toBe( true );
-			} );
-		} );
-
-		describe( 'lastBlockInserted()', () => {
-			it( 'should return client id of last block inserted', () => {
-				const expectedClientId = 1;
-				const action = {
-					type: 'ADD_LAST_BLOCK_INSERTED',
-					clientId: expectedClientId,
-				};
-
-				expect(
-					lastBlockInserted( { clientId: expectedClientId }, action )
-						.clientId
-				).toBe( expectedClientId );
-			} );
-
-			it( 'should return empty state if last block has been cleared', () => {
-				const action = {
-					type: 'CLEAR_LAST_BLOCK_INSERTED',
-				};
-
-				expect( lastBlockInserted( {}, action ) ).toStrictEqual( {} );
 			} );
 		} );
 	} );
